requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
et-xmlfile==1.1.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
openai~=0.28.1
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.230105
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0

langchain~=0.0.304